% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_data.R
\docType{data}
\name{airport_data}
\alias{airport_data}
\title{Latitude/Longitude Coordinates of different airports}
\format{
a dataframe containing all airports in the United States
\describe{
    \item{id}{name of the airport}
    \item{longtitude}{longtitude of the airport}
    \item{latitude}{latitude of the airport}
}
}
\usage{
airport_data
}
\description{
A dataset with the coordinates of all airports
}
\keyword{datasets}
