% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports_nearby.R
\name{airports_nearby}
\alias{airports_nearby}
\title{Returns the airport nearby a location given a distance and the longitude and latitude}
\usage{
airports_nearby(lonlat, airport_data, x = 75)
}
\arguments{
\item{lonlat}{is a character string that denotes ARP longitude}

\item{airport_data}{is the dataframe of the airport data}

\item{x}{is the distances nearby}
}
\value{
a character string list of nearby airports
}
\description{
Returns the airport nearby a location given a distance and the longitude and latitude
}
\examples{
airports_nearby(c(airport_data[airport_data$ID=="ITH", 2:3][[1]],
airport_data[airport_data$ID=="ITH", 2:3][[2]]), 
x=100, # chose Ithaca airport
airport_data)
}
