% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_nearby2.R
\name{airports_nearby2}
\alias{airports_nearby2}
\title{Returns the airport nearby a location given a distance and the longitude and latitude}
\usage{
airports_nearby2(lonlat, airport_data, x = 75)
}
\arguments{
\item{lonlat}{is a numeric string that denotes ARP longitude}

\item{airport_data}{is the dataframe of the airport data}

\item{x}{is the distances nearby}
}
\value{
a character string list of nearby airports
}
\description{
Returns the airport nearby a location given a distance and the longitude and latitude
}
