% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARPlat_to_deg.R
\name{ARPlat_to_deg}
\alias{ARPlat_to_deg}
\title{Converts a string character APR latitude to degree}
\usage{
ARPlat_to_deg(lat)
}
\arguments{
\item{lat}{is a character string that denotes ARP latitude}
}
\value{
deg of numeric
}
\description{
Converts a string character APR latitude to degree
}
\examples{
print(ARPlat_to_deg("040-39-08.5100N"))
}
