% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APRlon_to_deg.R
\name{ARPlon_to_deg}
\alias{ARPlon_to_deg}
\title{Converts a string character APR longtitude to degree}
\usage{
ARPlon_to_deg(lon)
}
\arguments{
\item{lon}{is a character string that denotes ARP longitude}
}
\value{
deg of numeric
}
\description{
Converts a string character APR longtitude to degree
}
\examples{
print(ARPlon_to_deg("075-26-25.4600W"))
}
