% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_symmetric.R
\name{check_symmetric}
\alias{check_symmetric}
\title{checks if a direct_from object is symmetric or not}
\usage{
check_symmetric(direct_from)
}
\arguments{
\item{direct_from}{air travel network list}
}
\value{
matrix with a boolean and a sub matrix
\itemize{
  \item \code{symm} TRUE or FALSE
  \item \code{conn} n x 2 matrix of asymmetric connections.
}
}
\description{
checks if a direct_from object is symmetric or not
}
\examples{
check_symmetric(direct_from)
}
