% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gcd.R
\name{compute_gcd}
\alias{compute_gcd}
\title{Calculates the distance between two lonlat distaces}
\usage{
compute_gcd(lonlat1, lonlat2)
}
\arguments{
\item{lonlat1}{is a matrix with longtitude and latitude in the form of string characters}

\item{lonlat2}{is a matrix with longtitude and latitude in the form of string characters}
}
\value{
distance of numeric
}
\description{
Calculates the distance between two lonlat distaces
}
\examples{
lonlat1 <- as.matrix(c("075-26-25.4600W","040-39-08.5100N"))
lonlat2 <- as.matrix(c("075-28-25.4600W","040-40-08.5100N"))
dist <- compute_gcd(lonlat1, lonlat2)
print(dist)
}
