% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gcd2.R
\name{compute_gcd2}
\alias{compute_gcd2}
\title{Calculates the distance between two lonlat distaces}
\usage{
compute_gcd2(lonlat1, lonlat2)
}
\arguments{
\item{lonlat1}{is a matrix with longtitude and latitude in the form numeric values}

\item{lonlat2}{is a matrix with longtitude and latitude in the form numeric values}
}
\value{
distance of numeric
}
\description{
Calculates the distance between two lonlat distaces
}
