% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_to_rad.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Degree to radian}
\usage{
deg_to_rad(degree)
}
\arguments{
\item{degree}{as a number}
}
\value{
radian
}
\description{
Given degree converts to radian
}
\examples{
print(deg_to_rad(180))
}
