% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_from.R
\docType{data}
\name{direct_from}
\alias{direct_from}
\title{Latitude/Longitude Coordinates of different airports location}
\format{
a list of airports with the information of direct flights from each of the airports
\describe{
    \item{id}{name of the airport}
}
}
\usage{
direct_from
}
\description{
A list of airports
}
\keyword{datasets}
