% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfun.R
\name{distfun}
\alias{distfun}
\title{Creates a matrix of distances between the two input matricies}
\usage{
distfun(lonlat1, lonlat2)
}
\arguments{
\item{lonlat1}{an n1 x 2 matrix, first column longitudes, second column latitudes}

\item{lonlat2}{an n2 x 2 matrix, first column longitudes, second column latitudes}
}
\value{
distmat - an n1 x n2 matrix of distances. distmat[i,j] is the distance between 
lonlat1[i,] and lonlat2[j,]
}
\description{
Creates a matrix of distances between the two input matricies
}
\examples{
airport_data_mat <- cbind(airport_data$longitude, airport_data$latitude)
rownames(airport_data_mat) <- airport_data$ID
colnames(airport_data_mat) <- c("longitude", "latitude")
round(distfun(airport_data_mat[c("BMI", "BOS", "ABE", "ERI"),], 
airport_data_mat[c("ALB","BOS","YUM"),]),3)
round(distfun(airport_data_mat, airport_data_mat)[1:10,1:10],3)
}
